banking-etl/requirements.txt
# ============================================================================
# Banking ETL Pipeline - Python Dependencies
# ============================================================================
# All required Python packages for the ETL pipeline
# Install with: pip install -r requirements.txt
# ============================================================================

# ============================================================================
# Core Data Processing
# ============================================================================
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# ============================================================================
# Database & ORM
# ============================================================================
SQLAlchemy==1.4.49
psycopg2-binary==2.9.7
alembic==1.12.0

# ============================================================================
# Apache Airflow & Plugins
# ============================================================================
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-docker==3.8.0
apache-airflow-providers-sqlite==3.4.0

# ============================================================================
# Big Data Processing (Optional for future scaling)
# ============================================================================
pyspark==3.4.1

# ============================================================================
# Logging & Monitoring
# ============================================================================
python-logging-loki==0.3.1
prometheus-client==0.17.1

# ============================================================================
# Testing & Code Quality
# ============================================================================
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pylint==3.0.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0
pydantic==1.10.13
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2

# ============================================================================
# Data Validation
# ============================================================================
great-expectations==0.17.12
